feat(render): use same html to text selectors and options to renderAsync and render
