+++ packages/render/src/optionts.ts
@@ -1,0 +1,7 @@
+import type { HtmlToTextOptions } from "html-to-text";
+
+export interface Options {
+  pretty?: boolean;
+  plainText?: boolean;
+  htmlToTextOptions?: HtmlToTextOptions;
+}
